refactor: update import paths in test files for consistency
