+++ frontend/src/components/__tests__/login.test.tsx
@@ -1,9 +1,9 @@
 import { describe, it, expect, beforeEach, vi, afterEach } from 'vitest';
 import { render, screen, fireEvent, waitFor } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import Login from '../frontend/src/components/login';
-import { api } from '../frontend/src/utils/api';
-import { getUserCache, clearUserCache } from '../frontend/src/utils/userCache';
+import Login from '../login';
+import { api } from '../../utils/api';
+import { getUserCache, clearUserCache } from '../../utils/userCache';
 
 // Mock the api module
 vi.mock('../../utils/api', () => ({
